update: list files recursively.

+++ skit_pipelines/components/upload2s3.py
@@ -1,5 +1,6 @@
 import kfp
 from kfp.components import InputPath
+from numpy import full
 
 from skit_pipelines import constants as pipeline_constants
 
@@ -16,9 +17,8 @@ def upload2s3(
 
     import boto3
     from loguru import logger
-
     from skit_pipelines.utils import create_file_name, create_dir_name
-
+    
     s3_resource = boto3.client("s3")
 
     if os.path.isfile(path_on_disk):
@@ -26,7 +26,9 @@ def upload2s3(
         s3_resource.upload_file(path_on_disk, bucket, )
     elif os.path.isdir(path_on_disk):
         upload_path = create_dir_name(org_id, file_type)
-        for full_file_path in glob(f"{path_on_disk}/**"):
+        for full_file_path in glob(f"{path_on_disk}/**", recursive=True):
+            if not os.path.isfile(full_file_path):
+                continue
             _, file_path = os.path.split(full_file_path)
             s3_resource.upload_file(full_file_path, bucket, os.path.join(upload_path, file_path))
